Add bitarray support; Migrage rcModeActivationMask to bitArray
