+++ src/main/blackbox/blackbox_fielddefs.h
@@ -111,6 +111,7 @@ typedef enum FlightLogEvent {
     FLIGHT_LOG_EVENT_SYNC_BEEP = 0,
     FLIGHT_LOG_EVENT_INFLIGHT_ADJUSTMENT = 13,
     FLIGHT_LOG_EVENT_LOGGING_RESUME = 14,
+    FLIGHT_LOG_EVENT_FLIGHTMODE = 30, // Add new event type for flight mode status.
     FLIGHT_LOG_EVENT_IMU_FAILURE = 40,
     FLIGHT_LOG_EVENT_LOG_END = 255
 } FlightLogEvent;
@@ -118,6 +119,11 @@ typedef enum FlightLogEvent {
 typedef struct flightLogEvent_syncBeep_s {
     uint32_t time;
 } flightLogEvent_syncBeep_t;
+
+typedef struct flightLogEvent_flightMode_s { // New Event Data type
+    uint32_t flags;
+    uint32_t lastFlags;
+} flightLogEvent_flightMode_t;
 
 typedef struct flightLogEvent_inflightAdjustment_s {
     uint8_t adjustmentFunction;
@@ -135,6 +141,7 @@ typedef struct flightLogEvent_loggingResume_s {
 
 typedef union flightLogEventData_u {
     flightLogEvent_syncBeep_t syncBeep;
+    flightLogEvent_flightMode_t flightMode; // New event data
     flightLogEvent_inflightAdjustment_t inflightAdjustment;
     flightLogEvent_loggingResume_t loggingResume;
 } flightLogEventData_t;
