+++ src/main/fc/cli.c
@@ -71,6 +71,7 @@ extern uint8_t __config_end;
 #include "fc/controlrate_profile.h"
 #include "fc/rc_adjustments.h"
 #include "fc/rc_controls.h"
+#include "fc/rc_modes.h"
 #include "fc/runtime_config.h"
 
 #include "flight/failsafe.h"
