+++ src/main/navigation/navigation.c
@@ -37,6 +37,7 @@
 #include "fc/fc_core.h"
 #include "fc/config.h"
 #include "fc/rc_controls.h"
+#include "fc/rc_modes.h"
 #include "fc/runtime_config.h"
 
 #include "flight/imu.h"
