+++ src/main/telemetry/telemetry.c
@@ -32,6 +32,7 @@
 
 #include "fc/config.h"
 #include "fc/rc_controls.h"
+#include "fc/rc_modes.h"
 #include "fc/runtime_config.h"
 
 #include "io/serial.h"
